requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.34.7
botocore==1.34.7
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
django-compat==1.0.15
jmespath==1.0.1
kombu==5.3.4
prompt-toolkit==3.0.43
pysrt==1.1.2
python-dateutil==2.8.2
python-magic==0.4.27
redis==5.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
